feat(shokunin): decoupling the metadata function
